Split the publishing work between the management and system servers
The portal reaches the management server, which runs management-api alone; the system
console reaches the system server, which runs both modules; and the management plane's one
path to the infrastructure is a proxy. So the records, their repositories and every service
the portal calls live in management-api, everything that touches nodes, workloads or Azure
lives in system-api, and DeploymentOperationsProxy, a @Proxy onto the system zone's new
DeploymentOperationsService, joins them.

Moved to system-api: OrganizationStorageProvisioner, OrganizationStorageService,
UiDeploymentProvisioner and UiStoragePaths with their Azure and mock implementations and
AzureUtil, the two property classes, now kinotic.systemApi.organizationStorage.* and
kinotic.systemApi.uiDeployment.* with the terraform, development, test and compose settings
renamed, the Azure SDK dependencies, and the organization provisioning job.

Moved to management-api: MicroserviceDeploymentService, published in its own zone with no
override. It, UiDeploymentService and DefaultOrganizationProvisioner authorize a request
against the caller's organization and delegate through the proxy: restart and remove a
microservice, check, provision and remove a site, provision an organization.
