+++ deployment/docker-compose/compose.kinotic-e2e-test.yml
@@ -30,17 +30,17 @@ services:
       KINOTIC_MANAGEMENTAPI_GITHUB_REPOTEMPLATE: "-"
       # Placeholder organization storage and UI site config, validated at boot the same way;
       # no e2e test deploys a UI, and both provisioners are the mocks.
-      KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_DISABLEPROVISIONER: "true"
-      KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_SUBSCRIPTIONIDS: "-"
-      KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_RESOURCEGROUP: "-"
-      KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_LOCATION: "-"
-      KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_PRIVATEENDPOINTSUBNETID: "-"
-      KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_PRIVATEDNSZONEID: "-"
-      KINOTIC_MANAGEMENTAPI_UIDEPLOYMENT_DISABLEPROVISIONER: "true"
-      KINOTIC_MANAGEMENTAPI_UIDEPLOYMENT_SITESDOMAIN: apps.localhost
-      KINOTIC_MANAGEMENTAPI_UIDEPLOYMENT_DNSZONEID: "-"
-      KINOTIC_MANAGEMENTAPI_UIDEPLOYMENT_FRONTDOORPROFILEID: "-"
-      KINOTIC_MANAGEMENTAPI_UIDEPLOYMENT_FRONTDOORENDPOINTHOSTNAME: "-"
+      KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_DISABLEPROVISIONER: "true"
+      KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_SUBSCRIPTIONIDS: "-"
+      KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_RESOURCEGROUP: "-"
+      KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_LOCATION: "-"
+      KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_PRIVATEENDPOINTSUBNETID: "-"
+      KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_PRIVATEDNSZONEID: "-"
+      KINOTIC_SYSTEMAPI_UIDEPLOYMENT_DISABLEPROVISIONER: "true"
+      KINOTIC_SYSTEMAPI_UIDEPLOYMENT_SITESDOMAIN: apps.localhost
+      KINOTIC_SYSTEMAPI_UIDEPLOYMENT_DNSZONEID: "-"
+      KINOTIC_SYSTEMAPI_UIDEPLOYMENT_FRONTDOORPROFILEID: "-"
+      KINOTIC_SYSTEMAPI_UIDEPLOYMENT_FRONTDOORENDPOINTHOSTNAME: "-"
       # DeploymentProperties.serverHost is @NotBlank and validated at boot. It is set in the
       # `development` profile, which e2e drops in favor of `test,e2e-tests,compose`, so supply it
       # here. No e2e test deploys a workload, so the value only has to bind.
