+++ deployment/terraform/azure/cluster/kinotic.tf
@@ -43,16 +43,16 @@ resource "helm_release" "kinotic_server" {
     { name = "extraEnv.KINOTIC_DOMAIN_SECRET_STORAGE_BACKEND", value = "azure" },
     { name = "extraEnv.KINOTIC_DOMAIN_SECRET_STORAGE_AZURE_VAULT_URL", value = azurerm_key_vault.main.vault_uri },
     # Organization storage — where kinotic-server provisions each organization's account
-    { name = "extraEnv.KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_SUBSCRIPTIONIDS", value = data.azurerm_client_config.keyvault.subscription_id },
-    { name = "extraEnv.KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_RESOURCEGROUP", value = azurerm_resource_group.org_storage.name },
-    { name = "extraEnv.KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_LOCATION", value = var.location },
-    { name = "extraEnv.KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_PRIVATEENDPOINTSUBNETID", value = azurerm_subnet.private_endpoints.id },
-    { name = "extraEnv.KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_PRIVATEDNSZONEID", value = azurerm_private_dns_zone.blob.id },
+    { name = "extraEnv.KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_SUBSCRIPTIONIDS", value = data.azurerm_client_config.keyvault.subscription_id },
+    { name = "extraEnv.KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_RESOURCEGROUP", value = azurerm_resource_group.org_storage.name },
+    { name = "extraEnv.KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_LOCATION", value = var.location },
+    { name = "extraEnv.KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_PRIVATEENDPOINTSUBNETID", value = azurerm_subnet.private_endpoints.id },
+    { name = "extraEnv.KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_PRIVATEDNSZONEID", value = azurerm_private_dns_zone.blob.id },
     # UI sites — the Front Door profile and DNS zone kinotic-server serves published UIs from
-    { name = "extraEnv.KINOTIC_MANAGEMENTAPI_UIDEPLOYMENT_SITESDOMAIN", value = local.sites_domain },
-    { name = "extraEnv.KINOTIC_MANAGEMENTAPI_UIDEPLOYMENT_DNSZONEID", value = local.global.dns_zone_id },
-    { name = "extraEnv.KINOTIC_MANAGEMENTAPI_UIDEPLOYMENT_FRONTDOORPROFILEID", value = azurerm_cdn_frontdoor_profile.sites.id },
-    { name = "extraEnv.KINOTIC_MANAGEMENTAPI_UIDEPLOYMENT_FRONTDOORENDPOINTHOSTNAME", value = azurerm_cdn_frontdoor_endpoint.sites.host_name },
+    { name = "extraEnv.KINOTIC_SYSTEMAPI_UIDEPLOYMENT_SITESDOMAIN", value = local.sites_domain },
+    { name = "extraEnv.KINOTIC_SYSTEMAPI_UIDEPLOYMENT_DNSZONEID", value = local.global.dns_zone_id },
+    { name = "extraEnv.KINOTIC_SYSTEMAPI_UIDEPLOYMENT_FRONTDOORPROFILEID", value = azurerm_cdn_frontdoor_profile.sites.id },
+    { name = "extraEnv.KINOTIC_SYSTEMAPI_UIDEPLOYMENT_FRONTDOORENDPOINTHOSTNAME", value = azurerm_cdn_frontdoor_endpoint.sites.host_name },
     # Email (Azure Communication Services) — shared service from global terraform
     { name = "extraEnv.KINOTIC_EMAIL_BACKEND", value = "azure" },
     { name = "extraEnv.KINOTIC_EMAIL_AZURE_ENDPOINT", value = local.global.email_service_endpoint },
