+++ kinotic-management-api/build.gradle
@@ -24,11 +24,6 @@ dependencies {
 
     implementation 'com.azure:azure-communication-email'
     implementation 'com.azure:azure-identity'
-    implementation 'com.azure:azure-storage-blob'
-    implementation 'com.azure.resourcemanager:azure-resourcemanager-cdn'
-    implementation 'com.azure.resourcemanager:azure-resourcemanager-dns'
-    implementation 'com.azure.resourcemanager:azure-resourcemanager-network'
-    implementation 'com.azure.resourcemanager:azure-resourcemanager-storage'
 
     implementation 'org.apache.ignite:ignite-core'
