+++ kinotic-system-api/build.gradle
@@ -11,6 +11,13 @@ dependencies {
 
     implementation 'io.projectreactor:reactor-core'
 
+    implementation 'com.azure:azure-identity'
+    implementation 'com.azure:azure-storage-blob'
+    implementation 'com.azure.resourcemanager:azure-resourcemanager-cdn'
+    implementation 'com.azure.resourcemanager:azure-resourcemanager-dns'
+    implementation 'com.azure.resourcemanager:azure-resourcemanager-network'
+    implementation 'com.azure.resourcemanager:azure-resourcemanager-storage'
+
     implementation 'io.vertx:vertx-core'
 
     implementation 'co.elastic.clients:elasticsearch-java'
