+++ kinotic-system-api/src/main/java/org/kinotic/system/api/config/SystemApiProperties.java
@@ -27,4 +27,16 @@ public class SystemApiProperties {
     @Valid
     private DeploymentProperties deployment = new DeploymentProperties();
 
+    /**
+     * Where each organization's storage account is provisioned.
+     */
+    @Valid
+    private OrganizationStorageProperties organizationStorage = new OrganizationStorageProperties();
+
+    /**
+     * Where each published UI is served from.
+     */
+    @Valid
+    private UiDeploymentProperties uiDeployment = new UiDeploymentProperties();
+
 }
