+++ website/content/02.platform/02.deployment-guide.md
@@ -50,7 +50,7 @@ kinotic-server identity the roles that takes on the resource group (Storage Acco
 Contributor, Storage Blob Data Contributor, Network Contributor), on the subnet (Network
 Contributor, to join it) and on the DNS zone (Private DNS Zone Contributor), and
 `kinotic.tf` hands the server the resource ids as
-`KINOTIC_MANAGEMENTAPI_ORGANIZATIONSTORAGE_*` environment variables (see
+`KINOTIC_SYSTEMAPI_ORGANIZATIONSTORAGE_*` environment variables (see
 [Configuration](/platform/configuration#organization-storage)).
 
 #### UI sites on Azure
@@ -61,6 +61,6 @@ organization an origin group, an origin on the organization's storage account an
 set, and per site a custom domain with a managed certificate, a route, and the CNAME and
 validation TXT records in the platform's DNS zone. `keyvault.tf` grants the kinotic-server
 identity CDN Profile Contributor on the profile and DNS Zone Contributor on the zone, and
-`kinotic.tf` passes the profile, endpoint and zone as `KINOTIC_MANAGEMENTAPI_UIDEPLOYMENT_*`
+`kinotic.tf` passes the profile, endpoint and zone as `KINOTIC_SYSTEMAPI_UIDEPLOYMENT_*`
 environment variables (see [Configuration](/platform/configuration#ui-sites)). The
 `kinotic-server` values file for development disables the provisioner instead.
