+++ website/content/02.platform/03.configuration.md
@@ -162,7 +162,7 @@ the runtime VMs.
 
 A deployment of a commit that contains a UI needs storage to publish it to. The platform
 keeps one storage account per organization, provisioned when the organization is created and
-configured under `kinotic.managementApi.organizationStorage.*`:
+configured under `kinotic.systemApi.organizationStorage.*`:
 
 | Property | Default | Meaning |
 |---|---|---|
@@ -184,7 +184,7 @@ provisioner and points at a local Azurite.
 ## UI sites
 
 Each published UI is served from its own site, a hostname under a platform domain fronted by
-Azure Front Door, configured under `kinotic.managementApi.uiDeployment.*`:
+Azure Front Door, configured under `kinotic.systemApi.uiDeployment.*`:
 
 | Property | Default | Meaning |
 |---|---|---|
