+++ website/content/02.platform/11.reference/03.project-publishing-design.md
@@ -29,6 +29,17 @@ ProjectDeployment            one per project: checkout dir, sync and publish VMs
 children are plain `AbstractRepository` entities in management-api beside `Workload`, not
 organization-scoped ones. Each child has a published service for the console with
 `findAllForProject` and `remove`; removal is the only path that destroys anything.
+
+Module placement follows the two servers the platform is splitting into: a management server
+running management-api alone, which the portal and every organization machine reach, and a
+system server running management-api and system-api, which the system console reaches. So
+the records, their repositories and every service the portal calls live in management-api;
+everything that touches nodes, workloads or Azure lives in system-api; and the management
+plane reaches the second only through `DeploymentOperationsProxy`, a `@Proxy` interface onto
+the system zone's `DeploymentOperationsService` (`restartMicroservice`, `removeMicroservice`,
+`checkUiSite`, `provisionUiSite`, `removeUiSite`, `provisionOrganization`). The management
+services authorize a request against the caller's organization, then delegate; the
+operations service trusts its callers.
 `ProjectDeployment` keeps the sync identity and loses `runtimeWorkloadId` and
 `runtimeMachineIdentityId`, which move to the per-microservice rows.
 
@@ -84,12 +95,13 @@ storage** task and a **Prepare Front Door** task, both idempotent, owned by the 
 and recorded on it as `provisioningJobRunId`. `OrganizationService.provision` runs every
 `OrganizationProvisioner` (a domain hook) on the organization: the signup flow calls it once
 the organization's record is complete, and `SystemOrganizationService.provisionOrganization`
-runs it again on request. The management module's provisioner starts the job and returns; the
-tasks record their outcome on the organization and the run shows in the system console's job
-runs and on the organization's overview, with a **Provision again** action. A deployment only
+runs it again on request. The management module's provisioner asks the system server to
+start the job through the proxy and returns; the tasks record their outcome on the
+organization and the run shows in the system console's job runs and on the organization's
+overview, with a **Provision again** action. A deployment only
 reads the outcome, and fails naming the state when the storage is not `READY`; nothing is
 provisioned by a deployment. A mock provisioner under
-`kinotic.managementApi.organizationStorage.disableProvisioner=true` points every organization
+`kinotic.systemApi.organizationStorage.disableProvisioner=true` points every organization
 at the configured Azurite connection string.
 
 ```
@@ -182,9 +194,9 @@ exported by `@kinotic-ai/core`. There is no push mechanism.
 
 ## Properties and dependencies
 
-`kinotic.managementApi.organizationStorage.*`: `subscriptionIds`, `resourceGroup`, `location`,
+`kinotic.systemApi.organizationStorage.*`: `subscriptionIds`, `resourceGroup`, `location`,
 `privateEndpointSubnetId`, `privateDnsZoneId`, `azuriteConnectionString`, `disableProvisioner`.
-`kinotic.managementApi.uiDeployment.*`: `sitesDomain`, `dnsZoneId`, `frontDoorProfileId`,
+`kinotic.systemApi.uiDeployment.*`: `sitesDomain`, `dnsZoneId`, `frontDoorProfileId`,
 `frontDoorEndpointHostName`, `disableProvisioner`. Nothing else is a property.
 
 Managed artifacts, each a `*Version` in `gradle.properties` and one line in the conventions
@@ -216,16 +228,17 @@ notification of publishes, and service endpoints instead of private endpoints.
   `microserviceDeployments`. `ProjectDeployment` no longer carries a runtime workload or
   identity; `findProjectMachines` lists the sync identity then one per microservice.
 - **Microservices in the console.** `MicroserviceDeploymentService` (`findAllForProject`,
-  `restart`, `remove`), published in the management-api zone from system-api, and the
-  portal's deployment page: a microservices table with logs, restart and remove, and the
+  `restart`, `remove`), published from management-api and reaching the VM through
+  `DeploymentOperationsProxy`, and the portal's deployment page: a microservices table with logs, restart and remove, and the
   machines labelled by the deployment that records them.
 - **Organization storage.** `AzureOrganizationStorageProvisioner` creates the account, the
   `ui` container, and the private endpoint with its DNS zone group, recording the outcome on
   `Organization` with a status of `PROVISIONING`, `READY` or `FAILED`. It is the first task
-  of the `provision-organization-<id>` job that `DefaultOrganizationProvisioner`, the
-  management module's `OrganizationProvisioner`, runs when the organization is created and
+  of the `provision-organization-<id>` job that `DeploymentOperationsService` runs on the
+  system server when the organization is created, asked through the proxy by
+  `DefaultOrganizationProvisioner`, the management module's `OrganizationProvisioner`, and
   whenever the system console's **Provision again** asks; the Front Door preparation is the
-  second. A deployment that publishes a UI reads the outcome and fails when the storage is
+  second. The provisioners, the storage service and their settings live in system-api. A deployment that publishes a UI reads the outcome and fails when the storage is
   not ready.
   `MockOrganizationStorageProvisioner` points every
   organization at Azurite. Terraform owns the resource group, private-endpoints subnet, private DNS zone and the
@@ -262,10 +275,11 @@ notification of publishes, and service endpoints instead of private endpoints.
   endpoint and grants the server CDN Profile Contributor and DNS Zone Contributor; the
   storage account's public network is open so Front Door can read it.
 - **UI deployments in the console.** `UiDeploymentService` (`findAllForProject`,
-  `retryProvisioning`, `remove`), published from management-api: listing checks any site
-  left provisioning for ten minutes or more, so one whose polling died with the server still
-  advances; retry runs the provisioner again and records the outcome; remove takes the site
-  down, deletes the UI's prefix in storage and deletes the row. Deleting a project removes
+  `retryProvisioning`, `remove`), published from management-api and delegating to the
+  system server through `DeploymentOperationsProxy`: listing has any site left provisioning
+  for ten minutes or more checked, so one whose polling died with the server still advances;
+  retry runs the provisioner again and records the outcome; remove takes the site down,
+  deletes the UI's prefix in storage and deletes the row. Deleting a project removes
   its UI deployments the same way. `@kinotic-ai/management-api` exposes it as
   `Kinotic.uiDeployments` with the `UiDeployment` model.
 - **Sites in the console, and the tab contract.** `UiDeployment.url` is minted with the
